增加JUMP模块 修正延迟函数的BUG

+++ SourceMINI58/Keil/Timer_Ctrl.c
@@ -32,7 +32,7 @@ uint32_t getSystemTime(void)						//系统时间，单位ms
 
 void DelayMsec(uint16_t time)
 {
-	uint16_t currentTime = getSystemTime();
+	uint32_t currentTime = getSystemTime();
 	while(getSystemTime() < (currentTime + time));
 }
